@@ -1,4 +1,4 @@
-<Router AppAssembly="@typeof(Program).Assembly" PreferExactMatches="@true"
+<Router AppAssembly="@typeof(Program).Assembly"
         AdditionalAssemblies="@(new [] {
             typeof(SoloX.BlazorJsonLocalization.Example.Components.StaticAssets.Pages.Counter).Assembly,
             typeof(SoloX.BlazorJsonLocalization.Example.Components.Embedded.Pages.Index).Assembly,
@@ -7,8 +7,10 @@
         })">
     <Found Context="routeData">
         <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
+        <FocusOnNavigate RouteData="@routeData" Selector="h1" />
     </Found>
     <NotFound>
+        <PageTitle>Not found</PageTitle>
         <LayoutView Layout="@typeof(MainLayout)">
             <p>Sorry, there's nothing at this address.</p>
         </LayoutView>
